gh-111386: Fix `uint32_t` cast in `generated_cases.c.h` (#111387)

+++ Python/bytecodes.c
@@ -2686,7 +2686,7 @@ dummy_func(
             assert(next_instr[oparg].op.code == END_FOR ||
                    next_instr[oparg].op.code == INSTRUMENTED_END_FOR);
             assert(1 + INLINE_CACHE_ENTRIES_FOR_ITER == next_instr - frame->instr_ptr);
-            frame->return_offset = 1 + INLINE_CACHE_ENTRIES_FOR_ITER + oparg;
+            frame->return_offset = (uint16_t)(1 + INLINE_CACHE_ENTRIES_FOR_ITER + oparg);
             DISPATCH_INLINED(gen_frame);
         }
